Made changed to TGAEngine so that it will compile properly. Added SDL.h platform independent include

+++ lib/TGAEngine/src/Animation.cpp
@@ -170,7 +170,7 @@ namespace TGA
 
 	GLuint Animation::getFrameCount()
 	{
-		return frames.size();
+		return (GLuint)frames.size();
 	}
 
 	std::string Animation::printFrames()
